Extract default page sources from CalendarPageSource.kt file
